These are all the files for Shopping List project on ZTM
The first script.js is how the javascript file started. The refracturedscript is the same script but refactored to run better.
The solved script is how I managed to get the script to all functionality.
The solved2 script is where I deleted the "done" button and made it work just by clicking the list item like the course wanted me to.